moved opportunity services test from translation test namespace to opportunity service test namespace. mocked opportunity service tests that were dependent on data that no longer existed in Db.
